finish moving to controllerutils
