+++ src/main/java/frc/util/Util.java
@@ -14,27 +14,6 @@ public class Util {
   }
 
   /**
-   * Scales radial deadband
-   *
-   * <p>The deadband value is set to be the new zero
-   *
-   * @param value the raw value
-   * @param deadband the deadband range
-   * @return the value with radial deadband applied
-   */
-  public static double deadband(double value, double deadband) {
-    if (Math.abs(value) > deadband) {
-      if (value > 0.0) {
-        return (value - deadband) / (1.0 - deadband);
-      } else {
-        return (value + deadband) / (1.0 - deadband);
-      }
-    } else {
-      return 0.0;
-    }
-  }
-
-  /**
    * This function finds the degree difference between angles, the shortest path. useful for pid
    * control of drivebase rotation
    *
